+++ aldrin/src/low_level.rs
@@ -5,9 +5,11 @@
 mod event;
 mod event_listener;
 mod proxy;
+mod reply;
 
 pub(crate) use event_listener::{EventListenerId, EventListenerRequest};
 
 pub use event::Event;
 pub use event_listener::EventListener;
 pub use proxy::Proxy;
+pub use reply::Reply;