+++ aldrin/src/reply.rs
@@ -1,7 +1,6 @@
-use crate::core::message::CallFunctionResult;
 use crate::core::Deserialize;
 use crate::error::Error;
-use futures_channel::oneshot::Receiver;
+use crate::low_level::Reply as LlReply;
 use std::fmt;
 use std::future::Future;
 use std::marker::PhantomData;
@@ -11,18 +10,26 @@ use std::task::{Context, Poll};
 /// Future to await the result of a call.
 #[must_use = "futures do nothing unless you `.await` or poll them"]
 pub struct Reply<T, E> {
-    recv: Receiver<Result<CallFunctionResult, Error>>,
-    function: u32,
+    inner: LlReply,
     phantom: PhantomData<fn() -> (T, E)>,
 }
 
 impl<T, E> Reply<T, E> {
-    pub(crate) fn new(recv: Receiver<Result<CallFunctionResult, Error>>, function: u32) -> Self {
+    pub(crate) fn new(inner: LlReply) -> Self {
         Self {
-            recv,
-            function,
+            inner,
             phantom: PhantomData,
         }
+    }
+
+    /// Casts the reply to resolve to different types.
+    pub fn cast<T2, E2>(self) -> Reply<T2, E2> {
+        Reply::new(self.inner)
+    }
+
+    /// Extracts the inner low-level reply.
+    pub fn into_low_level(self) -> crate::low_level::Reply {
+        self.inner
     }
 }
 
@@ -30,35 +37,18 @@ impl<T: Deserialize, E: Deserialize> Future for Reply<T, E> {
     type Output = Result<Result<T, E>, Error>;
 
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<Self::Output> {
-        match Pin::new(&mut self.recv).poll(cx) {
-            Poll::Ready(Ok(Ok(CallFunctionResult::Ok(t)))) => match t.deserialize() {
+        match Pin::new(&mut self.inner).poll(cx) {
+            Poll::Ready(Ok(Ok(t))) => match t.deserialize() {
                 Ok(t) => Poll::Ready(Ok(Ok(t))),
                 Err(e) => Poll::Ready(Err(Error::invalid_reply(e))),
             },
 
-            Poll::Ready(Ok(Ok(CallFunctionResult::Err(e)))) => match e.deserialize() {
+            Poll::Ready(Ok(Err(e))) => match e.deserialize() {
                 Ok(e) => Poll::Ready(Ok(Err(e))),
                 Err(e) => Poll::Ready(Err(Error::invalid_reply(e))),
             },
 
-            Poll::Ready(Ok(Ok(CallFunctionResult::Aborted))) => {
-                Poll::Ready(Err(Error::CallAborted))
-            }
-
-            Poll::Ready(Ok(Ok(CallFunctionResult::InvalidService))) => {
-                Poll::Ready(Err(Error::InvalidService))
-            }
-
-            Poll::Ready(Ok(Ok(CallFunctionResult::InvalidFunction))) => {
-                Poll::Ready(Err(Error::invalid_function(self.function)))
-            }
-
-            Poll::Ready(Ok(Ok(CallFunctionResult::InvalidArgs))) => {
-                Poll::Ready(Err(Error::invalid_arguments(self.function, None)))
-            }
-
-            Poll::Ready(Ok(Err(e))) => Poll::Ready(Err(e)),
-            Poll::Ready(Err(_)) => Poll::Ready(Err(Error::Shutdown)),
+            Poll::Ready(Err(e)) => Poll::Ready(Err(e)),
             Poll::Pending => Poll::Pending,
         }
     }
@@ -66,9 +56,6 @@ impl<T: Deserialize, E: Deserialize> Future for Reply<T, E> {
 
 impl<T, E> fmt::Debug for Reply<T, E> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("Reply")
-            .field("recv", &self.recv)
-            .field("function", &self.function)
-            .finish()
+        f.debug_struct("Reply").field("inner", &self.inner).finish()
     }
 }