+++ broker/src/broker/statistics/test.rs
@@ -201,9 +201,9 @@ async fn function_calls() {
     assert_eq!(stats.functions_replied, 0);
 
     // Call 2 functions.
-    let reply1 = proxy.call::<_, (), ()>(0, &());
+    let reply1 = proxy.call(0, &());
     let call1 = svc.next_function_call().await.unwrap();
-    let reply2 = proxy.call::<_, (), ()>(0, &());
+    let reply2 = proxy.call(0, &());
     let call2 = svc.next_function_call().await.unwrap();
     let stats = broker.take_statistics().await.unwrap();
     assert_eq!(stats.messages_sent, 2);